SkinScent/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
Pillow==10.0.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
razorpay==1.3.0
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
twilio==8.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
